seperate a star logic into a header file
